refactor: Split ViewMarkdownPanel into modular components
Move large ViewMarkdownPanel into smaller modules (MarkdownEditor,
MarkdownFloatingMenu, MarkdownPanelStates, hooks, plugins, styles,
types) to improve structure and maintainability.

Fix view markdown resolution to fall back to the workspace notes root
when an attached private workspace path is missing, and add test
TestViewMarkdownLinkReadsPrivateWorkspacePathFromWorkspaceNotes.
