+++ frontend/src/components/ViewMarkdownPanel/styles.ts
@@ -1,0 +1,256 @@
+import type { SystemStyleObject } from '@chakra-ui/react'
+
+export const markdownPanelBodySx: SystemStyleObject = {
+  '.tld-markdown-editor': {
+    '--basePageBg': 'var(--bg-canvas)',
+    '--baseBase': '#0f172a',
+    '--baseBgSubtle': '#111c31',
+    '--baseBg': '#152238',
+    '--baseBgHover': '#1a2b46',
+    '--baseBgActive': '#203456',
+    '--baseLine': 'rgba(148, 163, 184, 0.18)',
+    '--baseBorder': 'rgba(148, 163, 184, 0.22)',
+    '--baseBorderHover': 'rgba(148, 163, 184, 0.28)',
+    '--baseSolid': '#22385a',
+    '--baseSolidHover': '#2a4772',
+    '--baseText': '#dbe6f5',
+    '--baseTextContrast': '#f8fbff',
+    '--accentBase': '#10233d',
+    '--accentBgSubtle': '#123053',
+    '--accentBg': '#153c68',
+    '--accentBgHover': '#1a4a82',
+    '--accentBgActive': '#205798',
+    '--accentLine': 'rgba(96, 165, 250, 0.4)',
+    '--accentBorder': 'rgba(96, 165, 250, 0.45)',
+    '--accentBorderHover': 'rgba(96, 165, 250, 0.6)',
+    '--accentSolid': '#3b82f6',
+    '--accentSolidHover': '#60a5fa',
+    '--accentText': '#93c5fd',
+    '--accentTextContrast': '#eff6ff',
+    display: 'flex',
+    flexDirection: 'column',
+    width: '100%',
+    height: '100%',
+    minHeight: 0,
+    background: 'var(--bg-canvas)',
+    color: '#dbe6f5',
+  },
+  '.tld-markdown-toolbar': {
+    display: 'flex',
+    flexWrap: 'nowrap',
+    alignItems: 'center',
+    borderBottom: '1px solid rgba(148, 163, 184, 0.18)',
+    background: 'var(--basePageBg)',
+    color: '#dbe6f5',
+    paddingInline: '0.5rem',
+    minHeight: '48px',
+    flexShrink: 0,
+  },
+  '.tld-markdown-toolbar-formatting': {
+    display: 'flex',
+    flexWrap: 'nowrap',
+    alignItems: 'center',
+    flex: '0 1 auto',
+    minWidth: 0,
+    overflow: 'hidden',
+  },
+  '.tld-markdown-toolbar-spacer': {
+    flex: '1 1 auto',
+    minWidth: '0.5rem',
+  },
+  '.tld-markdown-toolbar-actions': {
+    marginLeft: 'auto',
+    paddingLeft: '0.5rem',
+    borderLeft: '1px solid rgba(148, 163, 184, 0.12)',
+    flexShrink: 0,
+  },
+  '.tld-markdown-toolbar-action': {
+    height: '30px',
+    minWidth: '30px',
+    width: '30px',
+    paddingInline: '0',
+    borderRadius: '0.55rem',
+    fontWeight: 600,
+    background: 'transparent',
+    color: '#dbe6f5',
+  },
+  '.tld-markdown-toolbar-action svg': {
+    width: '15px',
+    height: '15px',
+  },
+  '.tld-markdown-toolbar-action:hover': {
+    background: 'rgba(96, 165, 250, 0.12)',
+  },
+  '.tld-markdown-toolbar-action-save': {
+    background: 'rgba(59, 130, 246, 0.22)',
+    color: '#eff6ff',
+  },
+  '.tld-markdown-toolbar-action-save:hover': {
+    background: 'rgba(96, 165, 250, 0.3)',
+  },
+  '.tld-markdown-toolbar-action[data-disabled], .tld-markdown-toolbar-action:disabled': {
+    opacity: 0.4,
+    background: 'transparent',
+  },
+  '.tld-markdown-toolbar button': {
+    color: '#dbe6f5',
+  },
+  '.tld-markdown-toolbar button:hover': {
+    background: 'rgba(96, 165, 250, 0.14)',
+  },
+  '.tld-markdown-toolbar [data-state="on"]': {
+    background: 'rgba(59, 130, 246, 0.24)',
+    color: '#eff6ff',
+  },
+  '.tld-markdown-toolbar [disabled]': {
+    opacity: 0.35,
+  },
+  '.tld-markdown-editor .mdxeditor-root-contenteditable': {
+    position: 'relative',
+    display: 'flex',
+    flexDirection: 'column',
+    flex: '1 1 auto',
+    minHeight: 0,
+    background: 'var(--bg-canvas)',
+  },
+  '.tld-markdown-editor .mdxeditor-root-contenteditable > div': {
+    display: 'flex',
+    flexDirection: 'column',
+    flex: '1 1 auto',
+    minHeight: 0,
+  },
+  '.tld-markdown-editor__content': {
+    flex: '1 1 auto',
+    minHeight: 0,
+    width: '100%',
+    padding: '1.25rem 1.5rem 2rem',
+    fontSize: '0.95rem',
+    lineHeight: 1.75,
+    color: '#dbe6f5',
+    background: 'var(--bg-canvas)',
+    outline: 'none',
+    boxShadow: 'none',
+    overflowY: 'auto',
+  },
+  '.tld-markdown-editor__content p': {
+    color: '#dbe6f5',
+    marginBottom: '1rem',
+  },
+  '.tld-markdown-editor__content h1': {
+    color: '#f8fbff',
+    fontSize: '1.75em',
+    fontWeight: 700,
+    marginTop: '1.5rem',
+    marginBottom: '0.75rem',
+    lineHeight: 1.25,
+  },
+  '.tld-markdown-editor__content h2': {
+    color: '#f8fbff',
+    fontSize: '1.4em',
+    fontWeight: 700,
+    marginTop: '1.4rem',
+    marginBottom: '0.6rem',
+    lineHeight: 1.3,
+  },
+  '.tld-markdown-editor__content h3': {
+    color: '#f8fbff',
+    fontSize: '1.2em',
+    fontWeight: 600,
+    marginTop: '1.3rem',
+    marginBottom: '0.5rem',
+    lineHeight: 1.35,
+  },
+  '.tld-markdown-editor__content h4': {
+    color: '#f8fbff',
+    fontSize: '1.1em',
+    fontWeight: 600,
+    marginTop: '1.2rem',
+    marginBottom: '0.4rem',
+  },
+  '.tld-markdown-editor__content h5': {
+    color: '#f8fbff',
+    fontSize: '1.05em',
+    fontWeight: 600,
+    marginTop: '1.1rem',
+    marginBottom: '0.4rem',
+  },
+  '.tld-markdown-editor__content h6': {
+    color: '#f8fbff',
+    fontSize: '1em',
+    fontWeight: 600,
+    marginTop: '1rem',
+    marginBottom: '0.4rem',
+  },
+  '.tld-markdown-editor__content a': {
+    color: '#93c5fd',
+  },
+  '.tld-markdown-editor__content blockquote': {
+    borderLeft: '4px solid',
+    borderLeftColor: 'rgba(96, 165, 250, 0.45)',
+    color: '#cbd5e1',
+    paddingLeft: '1rem',
+    marginInline: '0',
+    marginBottom: '1rem',
+    fontStyle: 'italic',
+  },
+  '.tld-markdown-editor__content code': {
+    background: 'rgba(30, 41, 59, 0.9)',
+    color: '#bfdbfe',
+    padding: '0.2rem 0.4rem',
+    borderRadius: '0.25rem',
+    fontSize: '0.875em',
+  },
+  '.tld-markdown-editor__content pre': {
+    background: '#020817',
+    color: '#dbe6f5',
+    padding: '1rem',
+    borderRadius: '0.5rem',
+    overflowX: 'auto',
+    marginBottom: '1rem',
+  },
+  '.tld-markdown-editor__content pre code': {
+    background: 'transparent',
+    padding: 0,
+    borderRadius: 0,
+    color: 'inherit',
+    fontSize: 'inherit',
+  },
+  '.tld-mermaid-markdown-block': {
+    margin: '0.75rem 0 1rem',
+  },
+  '.tld-mermaid-markdown-block__summary': {
+    cursor: 'pointer',
+    fontFamily: 'ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace',
+    fontSize: '0.875em',
+    color: '#94a3b8',
+    whiteSpace: 'nowrap',
+  },
+  '.tld-mermaid-markdown-block__meta': {
+    marginLeft: '0.75rem',
+    color: '#64748b',
+  },
+  '.tld-mermaid-markdown-block--stale .tld-mermaid-markdown-block__meta': {
+    color: '#facc15',
+  },
+  '.tld-mermaid-markdown-block--other .tld-mermaid-markdown-block__meta': {
+    opacity: 0.72,
+  },
+  '.tld-mermaid-markdown-block__textarea': {
+    width: '100%',
+    minHeight: '8rem',
+    resize: 'vertical',
+    border: 0,
+    outline: 'none',
+    color: 'inherit',
+    background: 'transparent',
+    font: 'inherit',
+    lineHeight: 'inherit',
+  },
+  '.tld-markdown-editor__content ul, .tld-markdown-editor__content ol': {
+    paddingLeft: '1.5rem',
+    marginBottom: '1rem',
+  },
+  '.tld-markdown-editor__content li': {
+    marginBottom: '0.25rem',
+  },
+}
